# app/src/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Drum Sheet Manager API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/songs": {
      "get": {
        "tags": [
          "songs"
        ],
        "summary": "List all songs",
        "description": "Read songs from vw_all_songs with optional filtering.",
        "operationId": "get_all_songs_songs_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive search on artist_en or song_name_en.",
              "title": "Q"
            },
            "description": "Case-insensitive search on artist_en or song_name_en."
          },
          {
            "name": "in_master",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by master presence flag.",
              "title": "In Master"
            },
            "description": "Filter by master presence flag."
          },
          {
            "name": "audio_available",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by audio availability.",
              "title": "Audio Available"
            },
            "description": "Filter by audio availability."
          },
          {
            "name": "drum_sheet_available",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by drum sheet availability.",
              "title": "Drum Sheet Available"
            },
            "description": "Filter by drum sheet availability."
          },
          {
            "name": "source_available",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by source availability.",
              "title": "Source Available"
            },
            "description": "Filter by source availability."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SongRow"
                  },
                  "title": "Response Get All Songs Songs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/songs/count": {
      "get": {
        "tags": [
          "songs"
        ],
        "summary": "Count all songs",
        "description": "Get total count from vw_all_songs for pagination.",
        "operationId": "get_all_songs_count_songs_count_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive search on artist_en or song_name_en.",
              "title": "Q"
            },
            "description": "Case-insensitive search on artist_en or song_name_en."
          },
          {
            "name": "in_master",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by master presence flag.",
              "title": "In Master"
            },
            "description": "Filter by master presence flag."
          },
          {
            "name": "audio_available",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by audio availability.",
              "title": "Audio Available"
            },
            "description": "Filter by audio availability."
          },
          {
            "name": "drum_sheet_available",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by drum sheet availability.",
              "title": "Drum Sheet Available"
            },
            "description": "Filter by drum sheet availability."
          },
          {
            "name": "source_available",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by source availability.",
              "title": "Source Available"
            },
            "description": "Filter by source availability."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongCount"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/songs/playable": {
      "get": {
        "tags": [
          "songs"
        ],
        "summary": "List playable songs",
        "description": "Read songs from vw_playable_songs.",
        "operationId": "get_playable_songs_songs_playable_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive search on artist_en or song_name_en.",
              "title": "Q"
            },
            "description": "Case-insensitive search on artist_en or song_name_en."
          },
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive substring filter on genre.",
              "title": "Genre"
            },
            "description": "Case-insensitive substring filter on genre."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SongRow"
                  },
                  "title": "Response Get Playable Songs Songs Playable Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/songs/playable/count": {
      "get": {
        "tags": [
          "songs"
        ],
        "summary": "Count playable songs",
        "description": "Get total count from vw_playable_songs for pagination.",
        "operationId": "get_playable_songs_count_songs_playable_count_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive search on artist_en or song_name_en.",
              "title": "Q"
            },
            "description": "Case-insensitive search on artist_en or song_name_en."
          },
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive substring filter on genre.",
              "title": "Genre"
            },
            "description": "Case-insensitive substring filter on genre."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongCount"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/songs/unplayable": {
      "get": {
        "tags": [
          "songs"
        ],
        "summary": "List unplayable songs",
        "description": "Read songs from vw_unplayable_songs.",
        "operationId": "get_unplayable_songs_songs_unplayable_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive search on artist_en or song_name_en.",
              "title": "Q"
            },
            "description": "Case-insensitive search on artist_en or song_name_en."
          },
          {
            "name": "missing_audio",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter rows that are missing audio.",
              "title": "Missing Audio"
            },
            "description": "Filter rows that are missing audio."
          },
          {
            "name": "missing_drum_sheet",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter rows that are missing drum sheet.",
              "title": "Missing Drum Sheet"
            },
            "description": "Filter rows that are missing drum sheet."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SongRow"
                  },
                  "title": "Response Get Unplayable Songs Songs Unplayable Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/songs/unplayable/count": {
      "get": {
        "tags": [
          "songs"
        ],
        "summary": "Count unplayable songs",
        "description": "Get total count from vw_unplayable_songs for pagination.",
        "operationId": "get_unplayable_songs_count_songs_unplayable_count_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive search on artist_en or song_name_en.",
              "title": "Q"
            },
            "description": "Case-insensitive search on artist_en or song_name_en."
          },
          {
            "name": "missing_audio",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter rows that are missing audio.",
              "title": "Missing Audio"
            },
            "description": "Filter rows that are missing audio."
          },
          {
            "name": "missing_drum_sheet",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter rows that are missing drum sheet.",
              "title": "Missing Drum Sheet"
            },
            "description": "Filter rows that are missing drum sheet."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongCount"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/songs/recent/count": {
      "get": {
        "tags": [
          "songs"
        ],
        "summary": "Count recently updated songs",
        "description": "Get total count from vw_recently_updated_songs for pagination.",
        "operationId": "get_recently_updated_songs_count_songs_recent_count_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive search on artist_en or song_name_en.",
              "title": "Q"
            },
            "description": "Case-insensitive search on artist_en or song_name_en."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongCount"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SongCount": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "total"
        ],
        "title": "SongCount"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}